Improve dtkImaging integration
