+++ src/layers/CMakeLists.txt
@@ -31,10 +31,3 @@ add_subdirectory(legacy/medImageIO)
 add_subdirectory(legacy/medRegistration)
 
 add_subdirectory(legacy/medVtkInria)
-
-find_package(dtkImaging QUIET)
-
-if(dtkImaging_FOUND)
-    message("loading dtkImaging")
-    add_subdirectory(medImagingCompatibility)
-endif(dtkImaging_FOUND)
