+++ src/layers/medCore/medCore.cpp
@@ -33,7 +33,7 @@ namespace pluginManager
             medCore::morphomathOperation::initialize(realpath);
             medCore::maskImage::initialize(realpath);
             medCore::singleFilterOperation::initialize(realpath);
-            medCore::converter::initialize(realpath);
+            medCore::dataConverter::initialize(realpath);
         }
     }
 }
@@ -44,6 +44,7 @@ DTK_DEFINE_CONCEPT(medAbstractDiffusionModelEstimationProcess,diffusionModelEsti
 DTK_DEFINE_CONCEPT(medAbstractDiffusionScalarMapsProcess,diffusionScalarMaps)
 DTK_DEFINE_CONCEPT(medAbstractTractographyProcess,tractography)
 DTK_DEFINE_CONCEPT(medAbstractMaskImageProcess,maskImage)
+DTK_DEFINE_CONCEPT(medAbstractDataConverter,dataConverter)
 
 // arithimetic
 namespace arithmeticOperation
@@ -376,31 +377,6 @@ namespace morphomathOperation
         {
             return _private::factory;
         }
-    }
-}
-
-namespace converter
-{
-    namespace _private
-    {
-        medAbstractConverterPluginManager manager;
-        medAbstractConverterPluginFactory factory;
-    }
-
-    medAbstractConverterPluginManager& pluginManager(void)
-    {
-        return _private::manager;
-    }
-
-    medAbstractConverterPluginFactory& pluginFactory(void)
-    {
-        return _private::factory;
-    }
-
-    void initialize(const QString& path, bool verbose)
-    {
-        pluginManager().setVerboseLoading(verbose);
-        pluginManager().initialize(path);
     }
 }
