--- src/layers/medImagingCompatibility/dtkImagingCompatibilityExtensionPlugin.json
@@ -1,6 +1,0 @@
-{
-            "name" : "dtkImagingCompatibilityExtensionPlugin",
-         "concept" : "dtkComposerExtension",
-         "version" : "0.0.1",
-    "dependencies" : []
-}
