--- src/plugins/converter/converterPlugin.cpp
@@ -1,26 +1,0 @@
-#include "converter.h"
-#include "converterPlugin.h"
-
-#include <dtkCore>
-#include <medCore.h>
-
-// ///////////////////////////////////////////////////////////////////
-// XMLSaverPlugin
-// ///////////////////////////////////////////////////////////////////
-
-void converterPlugin::initialize(void)
-{
-    medCore::converter::pluginFactory().record("converterPlugin",converterCreator);
-}
-
-void converterPlugin::uninitialize(void)
-{
-
-}
-
-// ///////////////////////////////////////////////////////////////////
-// Plugin meta data
-// ///////////////////////////////////////////////////////////////////
-
-DTK_DEFINE_PLUGIN(converter)
-
