--- src/plugins/converter/converterPlugin.json
@@ -1,6 +1,0 @@
-{
-            "name" : "converterPlugin",
-         "concept" : "medAbstractConverter",
-         "version" : "0.0.1",
-    "dependencies" : []
-}
