--- src/plugins/converter/converterPlugin.qrc
@@ -1,5 +1,0 @@
-<RCC>
-    <qresource prefix="/">
-        <file>converterPlugin.json</file>
-    </qresource>
-</RCC>
