+++ src/plugins/medITKImageDTKImageConverter/CMakeLists.txt
@@ -11,18 +11,18 @@
 #
 ################################################################################
 
-set(TARGET_NAME medImagingCompatibility)
-
+set(TARGET_NAME medITKImageDTKImageConverterPlugin)
 
 ## #############################################################################
-## find requireds
+## Resolve dependencies
 ## #############################################################################
 
-find_package(dtk REQUIRED)
-include_directories(${dtk_INCLUDE_DIRS})
+find_package(ITK)
+include(${ITK_USE_FILE})
 
 find_package(dtkImaging REQUIRED)
 include_directories(${dtkImaging_INCLUDE_DIRS})
+
 
 ## #############################################################################
 ## List Sources
@@ -48,12 +48,13 @@ add_library(${TARGET_NAME} SHARED
 
 
 ## #############################################################################
-## include directorie.
+## include directories.
 ## #############################################################################
 
 target_include_directories(${TARGET_NAME}
-  PUBLIC ${${TARGET_NAME}_INCLUDE_DIRS}
+  PRIVATE ${${TARGET_NAME}_INCLUDE_DIRS}
   )
+# PRIVATE because plugins are not supposed to be used somewhere else
 
 
 ## #############################################################################
@@ -61,22 +62,19 @@ target_include_directories(${TARGET_NAME}
 ## #############################################################################
 
 target_link_libraries(${TARGET_NAME}
+  ITKCommon
   Qt5::Core
-  Qt5::Gui
-  medCore
-  dtkLog
-  dtkComposer
   dtkCore
+  dtkLog
+  medCore
+  medWidgets
   dtkImaging
-  # Legacy
-  medCoreLegacy
   )
 
 
 ## #############################################################################
-## install
+## Install rules
 ## #############################################################################
 
-set_plugin_install_rules(${TARGET_NAME}
-  ${${TARGET_NAME}_HEADERS}
-  )
+set_plugin_install_rules(${TARGET_NAME})
+
