+++ src/plugins/medITKImageDTKImageConverter/medITKImageDTKImageConverterPlugin.cpp
@@ -1,0 +1,22 @@
+#include <medITKImageDTKImageConverter.h>
+#include <medITKImageDTKImageConverterPlugin.h>
+
+#include <dtkCore>
+#include <medCore.h>
+
+void medITKImageDTKImageConverterPlugin::initialize(void)
+{
+    medCore::dataConverter::pluginFactory().record("medITKImageDTKImageConverterPlugin",medITKImageDTKImageConverterCreator);
+}
+
+void medITKImageDTKImageConverterPlugin::uninitialize(void)
+{
+
+}
+
+// ///////////////////////////////////////////////////////////////////
+// Plugin meta data
+// ///////////////////////////////////////////////////////////////////
+
+DTK_DEFINE_PLUGIN(medITKImageDTKImageConverter)
+
