+++ src/plugins/medITKImageDTKImageConverter/medITKImageDTKImageConverterPlugin.json
@@ -1,0 +1,6 @@
+{
+            "name" : "medITKImageDTKImageConverterPlugin",
+         "concept" : "medAbstractDataConverter",
+         "version" : "3.0.0",
+    "dependencies" : []
+}
