+++ src/plugins/medITKImageDTKImageConverter/medITKImageDTKImageConverterPlugin.qrc
@@ -1,0 +1,5 @@
+<RCC>
+    <qresource prefix="/">
+        <file>medITKImageDTKImageConverterPlugin.json</file>
+    </qresource>
+</RCC>
