+++ src/plugins/medImagingCompatibility/medImagingCompatibilityExtensionPlugin.cpp
@@ -1,15 +1,17 @@
-/* dtkImagingComposerExtension.cpp ---
- *
- * Author: Nicolas Niclausse
- * Copyright (C) Inria.
- */
-//
-//
+/*=========================================================================
 
+ medInria
 
-// Code:
+ Copyright (c) INRIA 2013 - 2014. All rights reserved.
+ See LICENSE.txt for details.
 
-#include "dtkImagingCompatibilityExtensionPlugin.h"
+  This software is distributed WITHOUT ANY WARRANTY; without even
+  the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+  PURPOSE.
+
+=========================================================================*/
+
+#include "medImagingCompatibilityExtensionPlugin.h"
 #include "medComposerImagingExtension.h"
 #include "dtkImaging.h"
 
@@ -20,27 +22,26 @@
 // Helper functions
 // ///////////////////////////////////////////////////////////////////
 
-dtkComposerExtension* dtkImagingCompatibilityCreator(void);
+dtkComposerExtension* medImagingCompatibilityCreator(void);
 
 // ///////////////////////////////////////////////////////////////////
 //
 // ///////////////////////////////////////////////////////////////////
 
 
-void dtkImagingCompatibilityExtensionPlugin::initialize(void)
+void medImagingCompatibilityExtensionPlugin::initialize(void)
 {
-    dtkComposer::extension::pluginFactory().record("dtkImagingCompatibility", dtkImagingCompatibilityCreator);
-    dtkComposerExtension *extension = dtkComposer::extension::pluginFactory().create("dtkImagingCompatibility");
+    dtkComposer::extension::pluginFactory().record("medImagingCompatibility", medImagingCompatibilityCreator);
+    dtkComposerExtension *extension = dtkComposer::extension::pluginFactory().create("medImagingCompatibility");
     bool verbose = dtkComposer::extension::pluginManager().verboseLoading();
     dtkImaging::setVerboseLoading(verbose);
     extension->extend(&(dtkComposer::node::factory()));
 }
 
-void dtkImagingCompatibilityExtensionPlugin::uninitialize(void)
+void medImagingCompatibilityExtensionPlugin::uninitialize(void)
 {
 
 }
-
 
 // ///////////////////////////////////////////////////////////////////
 // Plugin meta data
@@ -52,7 +53,7 @@ DTK_DEFINE_PLUGIN(dtkComposerExtension)
 // Helper functions
 // ///////////////////////////////////////////////////////////////////
 
-dtkComposerExtension *dtkImagingCompatibilityCreator(void)
+dtkComposerExtension *medImagingCompatibilityCreator(void)
 {
     return new medComposerImagingExtension();
 }
