+++ src/plugins/medImagingCompatibility/medImagingCompatibilityExtensionPlugin.json
@@ -1,0 +1,6 @@
+{
+            "name" : "medImagingCompatibilityExtensionPlugin",
+         "concept" : "dtkComposerExtension",
+         "version" : "3.0.0",
+    "dependencies" : []
+}
