+++ src/plugins/medImagingCompatibility/medToMedImageNode.json
@@ -1,7 +1,7 @@
 {
-          "title" : "To medImage converter",
+          "title" : "dtkImage To medImage converter",
            "kind" : "process",
-           "type" : "medAbstractConverter",
+           "type" : "medAbstractDataConverter",
            "tags" : ["converter"],
     "description" : "<p>Convert from dtkImage to medAbstractImageData</p>",
          "inputs" : ["dtkImage"],
