+++ superbuild/projects_modules/ITK.cmake
@@ -64,12 +64,12 @@ set(cmake_args
   -DCMAKE_SHARED_LINKER_FLAGS=${${ep}_shared_linker_flags}  
   -DCMAKE_INSTALL_PREFIX=<INSTALL_DIR>
   -DBUILD_SHARED_LIBS=${BUILD_SHARED_LIBS_${ep}}
-  -DBUILD_EXAMPLES=OFF
-  -DBUILD_TESTING=OFF
-  -DModule_ITKIOPhilipsREC=ON
-  -DModule_ITKReview=ON
-  -DModule_ITKVtkGlue=ON
-  -DVTK_DIR=${VTK_DIR}
+  -DBUILD_EXAMPLES:BOOL=OFF
+  -DBUILD_TESTING:BOOL=OFF
+  -DModule_ITKIOPhilipsREC:BOOL=ON
+  -DModule_ITKReview:BOOL=ON
+  -DModule_ITKVtkGlue:BOOL=ON
+  -DVTK_DIR:PATH=${VTK_DIR}
   )
 
 ## #############################################################################
